Add Tag component and support dot size
